trace/requirements.txt
fastapi==0.115.6
jinja2==3.1.4
python-multipart==0.0.19
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
argon2-cffi==23.1.0
itsdangerous==2.2.0
httpx==0.28.1
bleach==6.2.0
google-genai==0.3.0
exa-py==1.7.1
structlog==24.4.0
tenacity==9.0.0
